requirements.txt
streamlit>=1.28.0
folium>=0.14.0
streamlit-folium>=0.15.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pandas>=2.0.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
